Move RouteManager into a managers.py
